+++ components/event/event-api/src/test/java/uk/gov/justice/services/components/query/api/interceptors/EventApiInterceptorChainProviderTest.java
@@ -7,14 +7,12 @@ import static uk.gov.justice.services.core.annotation.Component.EVENT_API;
 
 import uk.gov.justice.services.core.accesscontrol.LocalAccessControlInterceptor;
 import uk.gov.justice.services.core.audit.LocalAuditInterceptor;
-import uk.gov.justice.services.core.interceptor.Interceptor;
+import uk.gov.justice.services.core.interceptor.PriorityInterceptorType;
 import uk.gov.justice.services.metrics.interceptor.IndividualActionMetricsInterceptor;
 import uk.gov.justice.services.metrics.interceptor.TotalActionMetricsInterceptor;
 
 import java.util.List;
 
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Test;
 
 public class EventApiInterceptorChainProviderTest {
@@ -27,12 +25,12 @@ public class EventApiInterceptorChainProviderTest {
     @Test
     @SuppressWarnings("unchecked")
     public void shouldProvideDefaultInterceptorChainTypes() throws Exception {
-        final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new EventApiInterceptorChainProvider().interceptorChainTypes();
+        final List<PriorityInterceptorType> interceptorChainTypes = new EventApiInterceptorChainProvider().interceptorChainTypes();
 
         assertThat(interceptorChainTypes, containsInAnyOrder(
-                new ImmutablePair<>(1, TotalActionMetricsInterceptor.class),
-                new ImmutablePair<>(2, IndividualActionMetricsInterceptor.class),
-                new ImmutablePair<>(3000, LocalAuditInterceptor.class),
-                new ImmutablePair<>(4000, LocalAccessControlInterceptor.class)));
+                new PriorityInterceptorType(1, TotalActionMetricsInterceptor.class),
+                new PriorityInterceptorType(2, IndividualActionMetricsInterceptor.class),
+                new PriorityInterceptorType(3000, LocalAuditInterceptor.class),
+                new PriorityInterceptorType(4000, LocalAccessControlInterceptor.class)));
     }
 }