+++ components/event/event-listener/src/test/java/uk/gov/justice/services/components/event/listener/interceptors/it/EventBufferAndFilterChainIT.java
@@ -35,20 +35,20 @@ import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
 import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
 import uk.gov.justice.services.core.envelope.EnvelopeValidationExceptionHandlerProducer;
 import uk.gov.justice.services.core.enveloper.Enveloper;
-import uk.gov.justice.services.event.buffer.api.AbstractEventFilter;
-import uk.gov.justice.services.event.buffer.api.AllowAllEventFilter;
-import uk.gov.justice.services.core.extension.ServiceComponentScanner;
 import uk.gov.justice.services.core.extension.BeanInstantiater;
-import uk.gov.justice.services.core.interceptor.Interceptor;
+import uk.gov.justice.services.core.extension.ServiceComponentScanner;
 import uk.gov.justice.services.core.interceptor.InterceptorCache;
 import uk.gov.justice.services.core.interceptor.InterceptorChainObserver;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessorProducer;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProvider;
+import uk.gov.justice.services.core.interceptor.PriorityInterceptorType;
 import uk.gov.justice.services.core.json.DefaultJsonSchemaValidator;
 import uk.gov.justice.services.core.json.JsonSchemaLoader;
 import uk.gov.justice.services.core.requester.RequesterProducer;
 import uk.gov.justice.services.core.sender.SenderProducer;
+import uk.gov.justice.services.event.buffer.api.AbstractEventFilter;
+import uk.gov.justice.services.event.buffer.api.AllowAllEventFilter;
 import uk.gov.justice.services.event.buffer.core.service.ConsecutiveEventBufferService;
 import uk.gov.justice.services.messaging.DefaultJsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.JsonEnvelope;
@@ -71,8 +71,6 @@ import javax.sql.DataSource;
 import liquibase.Liquibase;
 import liquibase.database.jvm.JdbcConnection;
 import liquibase.resource.ClassLoaderResourceAccessor;
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.openejb.jee.Application;
 import org.apache.openejb.jee.WebApp;
 import org.apache.openejb.junit.ApplicationComposer;
@@ -248,10 +246,10 @@ public class EventBufferAndFilterChainIT {
         }
 
         @Override
-        public List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes() {
-            final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new ArrayList<>();
-            interceptorChainTypes.add(new ImmutablePair<>(1, EventBufferInterceptor.class));
-            interceptorChainTypes.add(new ImmutablePair<>(2, EventFilterInterceptor.class));
+        public List<PriorityInterceptorType> interceptorChainTypes() {
+            final List<PriorityInterceptorType> interceptorChainTypes = new ArrayList<>();
+            interceptorChainTypes.add(new PriorityInterceptorType(1, EventBufferInterceptor.class));
+            interceptorChainTypes.add(new PriorityInterceptorType(2, EventFilterInterceptor.class));
             return interceptorChainTypes;
         }
     }