+++ components/query/query-controller/src/test/java/uk/gov/justice/services/components/command/api/interceptors/QueryControllerInterceptorChainProviderTest.java
@@ -3,20 +3,17 @@ package uk.gov.justice.services.components.command.api.interceptors;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
-import static uk.gov.justice.services.core.annotation.Component.QUERY_API;
 import static uk.gov.justice.services.core.annotation.Component.QUERY_CONTROLLER;
 
 import uk.gov.justice.services.components.query.controller.interceptors.QueryControllerInterceptorChainProvider;
 import uk.gov.justice.services.core.accesscontrol.LocalAccessControlInterceptor;
 import uk.gov.justice.services.core.audit.LocalAuditInterceptor;
-import uk.gov.justice.services.core.interceptor.Interceptor;
+import uk.gov.justice.services.core.interceptor.PriorityInterceptorType;
 import uk.gov.justice.services.metrics.interceptor.IndividualActionMetricsInterceptor;
 import uk.gov.justice.services.metrics.interceptor.TotalActionMetricsInterceptor;
 
 import java.util.List;
 
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Test;
 
 public class QueryControllerInterceptorChainProviderTest {
@@ -29,13 +26,13 @@ public class QueryControllerInterceptorChainProviderTest {
     @Test
     @SuppressWarnings("unchecked")
     public void shouldProvideDefaultInterceptorChainTypes() throws Exception {
-        final List<Pair<Integer, Class<? extends Interceptor>>>
+        final List<PriorityInterceptorType>
                 interceptorChainTypes = new QueryControllerInterceptorChainProvider().interceptorChainTypes();
 
         assertThat(interceptorChainTypes, containsInAnyOrder(
-                new ImmutablePair<>(1, TotalActionMetricsInterceptor.class),
-                new ImmutablePair<>(2, IndividualActionMetricsInterceptor.class),
-                new ImmutablePair<>(3000, LocalAuditInterceptor.class),
-                new ImmutablePair<>(4000, LocalAccessControlInterceptor.class)));
+                new PriorityInterceptorType(1, TotalActionMetricsInterceptor.class),
+                new PriorityInterceptorType(2, IndividualActionMetricsInterceptor.class),
+                new PriorityInterceptorType(3000, LocalAuditInterceptor.class),
+                new PriorityInterceptorType(4000, LocalAccessControlInterceptor.class)));
     }
 }