+++ core/src/test/java/uk/gov/justice/services/core/interceptor/InterceptorCacheTest.java
@@ -18,8 +18,6 @@ import java.util.List;
 
 import javax.enterprise.inject.spi.Bean;
 
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -162,9 +160,9 @@ public class InterceptorCacheTest {
             }
 
             @Override
-            public List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes() {
-                final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new ArrayList<>();
-                interceptorChainTypes.add(new ImmutablePair<>(1, interceptorClass_1));
+            public List<PriorityInterceptorType> interceptorChainTypes() {
+                final List<PriorityInterceptorType> interceptorChainTypes = new ArrayList<>();
+                interceptorChainTypes.add(new PriorityInterceptorType(1, interceptorClass_1));
                 return interceptorChainTypes;
             }
         });
@@ -203,7 +201,7 @@ public class InterceptorCacheTest {
         }
 
         @Override
-        public List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes() {
+        public List<PriorityInterceptorType> interceptorChainTypes() {
             return emptyList();
         }
     }
@@ -216,10 +214,10 @@ public class InterceptorCacheTest {
         }
 
         @Override
-        public List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes() {
-            final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new ArrayList<>();
-            interceptorChainTypes.add(new ImmutablePair<>(1, InterceptorOne.class));
-            interceptorChainTypes.add(new ImmutablePair<>(3, InterceptorTwo.class));
+        public List<PriorityInterceptorType> interceptorChainTypes() {
+            final List<PriorityInterceptorType> interceptorChainTypes = new ArrayList<>();
+            interceptorChainTypes.add(new PriorityInterceptorType(1, InterceptorOne.class));
+            interceptorChainTypes.add(new PriorityInterceptorType(3, InterceptorTwo.class));
             return interceptorChainTypes;
         }
     }
@@ -232,10 +230,10 @@ public class InterceptorCacheTest {
         }
 
         @Override
-        public List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes() {
-            final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new ArrayList<>();
-            interceptorChainTypes.add(new ImmutablePair<>(1, InterceptorTwo.class));
-            interceptorChainTypes.add(new ImmutablePair<>(2, InterceptorThree.class));
+        public List<PriorityInterceptorType> interceptorChainTypes() {
+            final List<PriorityInterceptorType> interceptorChainTypes = new ArrayList<>();
+            interceptorChainTypes.add(new PriorityInterceptorType(1, InterceptorTwo.class));
+            interceptorChainTypes.add(new PriorityInterceptorType(2, InterceptorThree.class));
             return interceptorChainTypes;
         }
     }
@@ -248,9 +246,9 @@ public class InterceptorCacheTest {
         }
 
         @Override
-        public List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes() {
-            final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new ArrayList<>();
-            interceptorChainTypes.add(new ImmutablePair<>(2, InterceptorThree.class));
+        public List<PriorityInterceptorType> interceptorChainTypes() {
+            final List<PriorityInterceptorType> interceptorChainTypes = new ArrayList<>();
+            interceptorChainTypes.add(new PriorityInterceptorType(2, InterceptorThree.class));
             return interceptorChainTypes;
         }
     }